requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atom==0.11.0
backports.tarfile==1.2.0
blinker==1.9.0
bugsnag==4.7.1
build==1.2.2.post1
CacheControl==0.14.3
cerebrium==1.40.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==45.0.4
distlib==0.3.9
dulwich==0.22.8
exceptiongroup==1.3.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.13.1
findpython==0.6.3
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2024.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.3
keyring==25.6.0
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numpy==2.1.2
onnx==1.18.0
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
packaging==24.2
pbs-installer==2025.6.6
pillow==11.0.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
scikit-image==0.25.2
scipy==1.15.3
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tenacity==9.1.2
termcolor==2.3.0
tifffile==2025.5.10
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.3
virtualenv==20.31.2
WebOb==1.8.9
Werkzeug==3.1.3
yaspin==3.1.0
zipp==3.23.0
zstandard==0.23.0
